DINAJ trim support, left and right side monsterclock meters
DINAJ is now moved to top level and dinajT.lua is added which returns a
trimmed version of the info. Monsterclock folder format changed and now
incldues left and right side varients and every varient having every
option. May look into seeing if all varients can be displayed in folder
form to make changing easier if not the installer will take care of that
when its done anyways.