Fixed Ping graph timing bug.
Seems that the ping measure for the graph wants the udapte rate to be
defined as "UpdateRate" which I remember seeing in documentation that
ping and regex can use that. However if you declare "UpdateDivider"
after that it cause them to use that value instead. Thus I have removed
"UpdateDivider" from all and properly follow documentation.