Testing out TopProcesses meters
Odd bug where after the first two cpu sages that are not unique and only
the first name is unique gonna have to look up a meter that uses
advanced cpu as its manual does not give an example of getting second
usage and I must be using it wrong.
I will be doing tests to see if AdvancedCPU or NomFerp is less of a CPU
hog and I may still have advancedCPU as an option as I have heard people
say NomFerp is not supported on all systems.