Network and processing combined are now also dynamic color
Turns out DynamicVariables are only checked at update time, so on a slow
meter it takes several seconds (Also the variables are checked early on
so you can have it be almost 2x times the update time before it
updates), I guess now I should look into hot the redraw command works,
if it just redraws everything with news colors and doesnt impact update
rates or old info I may attach a redraw group to the EQ and use it as
sort of a redraw push mechanism. Also I need to get dynamic secondary
and background colors that match.