Dynamic color changes added to CPU small meters
I lied, I decided after being busy with life for a few days to go on
ahead and make the changes to the system anyways and just tag a warning
that it will not all update at once. Also removed bug where dynamic
wallpaper took up a core of CPU in last update.