Dynamic color V1
All meters now accept the color variable being changed. Going to rewrite
the color system to generate color variations for each of the color
options and bring back the default color working.

Also going to be looking into CPU usage of this feature and make sure
when it is off it is not causing extra strain. I am curious how much
dynamic variables impact CPU usage.