Changes to dynamic system
Turns out background cannot be dynamic unless they are outside the
rainmeter section, seeing as I have no way of knowing size of meter
without a lot of reconstruction to make background be the right height
background will not be dynamic. Turning on transparency though gives the
same effect so it is not a big deal.