CPU issues?
I forgot to mention that CPU is now back to normal 10%-15%, not sure
what cause that jump, I think my changes to the variationer to cover
some odd cases just fixed it is my guess. Also dynamic dynamic variables
is either really easy or is not working and I am not sure which it is.