Fact meter improvements
Fact meter now uses its own text size and I removed the fact not found
extras I had used to seperate the fact sections.

I have decided to delay or scrap the change to variants unless there is
a simple way to do it and include the alts without changing position and
even then I dont like it because of the number of alts I may end up
with. I may just replace it with maybe making the alts folder visible
under variants but I doubt you can do that.