IO Top Process
I started working on a top disk IO process meter, there is an issue at
the moment though where I have nothing to set as a max value of the
meters since I dont know disk usage max potential like with the CPU and
RAM meters. No idea if I will merge it with combined since it has that
issue. Also does not have alts yet

Updated README, I have decided not to add more info to CPU and GPU for
the time being, maybe if people request it a bunch I will as an alt.
However I would need to label all info the meter if that was the case
and I am not sure what 2 or 4 things I would add.