Changes to the complementary color system
This ended up being way less cool than I expected it to be, I may tweak
it some more to make it better but for some colors it just doesnt work
well, maybe in the future I will rewrite it to move L dynamically as
well but for now I am just leaving it as is with a giant warning. Also
made it no long split compiment colors as those really did not work
well, at least the current system looks good sometimes.