Investigating some graphs at 100% being lower than they should be
See CPU graph alt and alt2 for example. It seems the one is set to 0R
and one set to 0 but the 0 one looks better, I am trying some move
manual fixes but not sure if this is an intended thing with histograms
where they are given more room than the max value.