Auto album blur test
Currently uses imagemagick to run the script has a few current issues:

1. Between when the image changes that spotify has obtained and when the
song changes there is several seconds, the script currently runs right
after the song has changed.
2. The command window shows up for a moment, I will liekly need to
switch to a bat file to get around this.
3. Since the colors script is called for every meter it gets called way
more than it should. I need to move this script to just being attached
to the background meter and somehow get an alert when the file changes,
(Not sure if rainmeter has a filing system)
4. The script is called for when the song changes in other players not
spotify
5.I am not sure how to get the album art saved as a file for other
players.

BUT it does work at the moment and looks nice, likely will allow users
to configure blur strength in the future. I also may allow it to output
and monitor resolution but I doubt that will have an impact on how nice
it looks.