Updated logos and changes to album blur system
I wont be doing as many commits for the next couple days due to a
combination of being sick and busy. Also the fact that album art blur is
going to be a hassle to keep CPU usage down and get things passing
right.

As for album art blur in an attempt to keep CPU usage down the blur pass
will only be done on the albumArt backgrounds and thus I have moved it
to there. I found a bug where it seems to not run on load and only on
change.

Once it is fully functional I will then add the palette pass to happen
only when the albumart background is active since I need to reduce the
number of calls that it goes through before adding it to everything.