Album blur functional again
I have moved the blur function to a bat file that I will later move into
the resources folder with everything else that needs to be moved there.
I am hoping I can get this bat to run windowless but current atempts
seem to not be running the scripts I make. Once that is done I can move
on to fixing the current quirks with the blur system and move on to the
palette system.