Album blur finished
Thats right, finished. No cmd pop ups and everything works
amazinglyestism, yes that is a word I just made it up.

Palette system framework is now in place the way I want it to be expect
updates with that shortly, as to how I am going to go about it I am
still looking around to see if I can pull something off the shelf that
will be faster than lua scrips.