vbs script no running, no info returned
Working to get info out of it. between this version and the last version
something I did made a bunch of operable and another file that is non us
unicode that I had to wipe out, if I have not already fixed it it will
be fixed for the next release.