Both test and output files now properly output results of the command to file
Still hoping I can find a way to pass it back to lua without the file
interaction.

On a side note a list of 3000+ facts has been added to the list of
facts. As for the validity of these facts who knows. Also I noticed all
' were replaced with ? so likely I need to reget the file and save it
under a different coding or do some smart regexing to fit them.