Fixed an issue where both album art background for spotify did not work
Apparently the one did not get udpated. Also now waittime is a variable
able to be defined in the variables.ini.
Next update should have the delay built in to every color option and I
am debating if I want to have the get average color options run early as
well so that everything is done and the moment the wait time happens
everything updates.