Fix issue with clock
Fixed an issue where the clock would move left if total text got
smaller, for now it will just grow in size and stay that big not sure if
I want to keep it that way or always make it dynamic. Updated where I
had changed the variables file for me from what it should be. I am now
fairly certain that the output issue is not my code but iTunes going to
try on another machine tomorrow.