Split wait times.
Now there are two different wait timers, the first if intended to
accommodate the delay in fetching spotify album art and the second is to
accommodate the delay in processing images. I have noticed that if the
second delay is too long some meters will update early so I may build a
check into it to make it so if the process is done it will update early.

Also it turns out that the output files have been working all along and
I just had open a file I wasn't using. I am fairly certain that my
timing changes fixed it, going to be handling base cases and the like
shortly.