Color palette selection improvements
No bug fixes in this patch and I have not worked on the new threading
system. However I have noticed I am getting less program still running
errors (Just the occasional error where I try to kill it when it is not
running and my guess is that is because I try to kill it just as it
finishes)

Color palette now sorts the top several colors so that darker colors are
earlier on in the color list which should make color coding easier. I
have also added a palette meter to make visualizing easier and this will
be the first skin to use the new threading tests when I work on those. I
still want to polish the color selection a little more before I move on.