First pass of new implementation
Completely untested and does not output to files yet, next update should
hopefully be functional. A lot of refactoring going on so that edge
casses are better caught and logged and the code is much cleaner now.
Once this change is done I will get rid of unused script files and I
think I will move everything into the right place in resources