Fixed a bug where album art blur didnt run
Now fallback uses its own runner just in case, I have now seen a few
cases where threads can not be terrminated (I am guessing because they
are in the middle of fileoutput) but they still seem to run just fine (I
am guessing becasue there is enough time that when I go to run a command
they are dead.

1 last bug to go then it will just be removing that hacky log spam for
something else. The only bug that remains is that the first pass of both
fallback and normal output nothing and I think I know the cause it is
because the kill is being applied right after they run, if this is true
then me killing nonexistant threads may cause issues when that happens,
ill look into it.