Fixed first load bug for album art blur
I finally cracked it, that was a pain in my ass. Turns out the first run
instantly fails but runs almost like as if it is trying to init itself.
I hope I can use this fix on the song info, I think I can since the
write is on a delay and the read is not.