0.6.10 RC1
Going to test this release for a bit to see if there are any bugs that
have been missed.

Note this RC still has the spotify log spam and the time wasting log
spam. I have yet to find a way to reliably wait only 10ms in the lua
that comes with rainmeter.