First pass date check
CPU usage seems to be high on load, ill look into why. AlbumArt always
seems to end up being the fallback image I think because it is a higher
res image than the album art and overrides it in the end (Which my guess
is also why the CPU is high) I need to downsample all of them and see if
I can find a way to see if it really is nil because we are inbetween
songs or because there is no song.