Updated lower res images
Found a bug where the blurred image of the album art would sometimes
become the fallback art. I am looking into a workaround to fix it but
this bug may be around for a while. Also looking into ways to make the
spotify stuff pull on first load since it seems to not and to make it
run the fallback on close.