GPMDP first pass
Now functionally reads the album art, artist, and song info. Does not
yet pass these along to be used to make a file and I am pretty sure
randomly reload. Also duration does not work yet as it seems to be a
percent, I may rework all duration meters into a bar style so they are
all the same if I can not do websockets in rainmeter. Speaking of that
this is as good as support will get for features unless I can get
websockets to work.