Rewrote GPMDP lua file
The old one was a mess and more complex than needed, so I rebuilt it
using regular expressions, I mean patterns. Should now imitate much more
closely the original skins. Position and duration do exist in the
playback json so that are in now but are not formated right. Also album
art is not integrated but I should have that fixed shortly. I plan on
reworking the album art go it also sends a notify out to its group so
that way the album art doesn't try to download multiple times and
hopefully this should fix the delayed download bug.