Fixes to sleep system
Turns out the sleep was not actually doing anything but the length of
time that it took to do that was enough to solve the problem most of the
time. I have abused this knowledge to fix all but one of the remaining
problems with GPMDP I believe. The only remaining issue I believe is the
first run of the background blurring (And potentially the refresh
causing the colors to go black still).