String clipping added to top processes
I dont think it will need to be added to the other minibars so I have
left it off of that but if I need to change that in the future I will.
Git seems to be misdisplaying the changes for these files though so I
hope it will commit right, I just double checked the files and they are
all correct, I guess we will find out.