Fixed a bug where username couldnt have spaces
I have not fixed it yet with the bundled imagemagick, wanted to get to a
working base as this one was a pain to debug and I think that was
contributing to the problem as well. Should have a fix for that soon.
Also note I am away from my usual setup so don't expect anything but bug
fixes I am experiencing for the next 2ish weeks.