Verbosity system partial implementation
I have started implementing the verbosity system, so far it is
implmented on CPU GPU and system combined meters. Gonna be a slow
process though so it will take so time to do. Still looking into GPU
alternatives for laptops, the ones I have found are outdated or a lot of
manual work though.