Fixes to monsterclocks
My regex I used to replace the script to the new location was bad so
that has been fixed now. Also fixed a few random double returns that
were happening. Seems to be a few bugs when using not GPMDP that I have
introduced now, not sure if they are issues with a different using a
different rig or something else. Also updated versioning to reflect
current version. I think I will try to do a release with 0.7.7