Better music switcher
Trying to get the music switcher to work so that it reflects initial
state and changes made outside of it, while it internally reflects that
it is not changing like it should and I am getting a calc error. I wanna
get this fixed before I make the switcher switch everything else.