Complete music switcher
Music switcher now changes all music sources to new source.

I seem to have found an issue where you can cause rainmeter to crash if
you switch right away, I need to check if the new code is the cause of
this or if it is one of the old ways I squashed this bug doesn't work if
the source thinks it is starting from scratch again.