Initial transition to BetterMusicPlugin
Colors are not always updating, my guess is that the album art is
sometimes downloading too fast and thus the source is never changed but
that is an issue with BetterMusic/Cover download timing that I can work
at.
Performance is much better now and is around 3-5% with my usual full kit
of readouts. I also tweaked a few things with some of the older widgets
to make them more efficient, however I may have to increase the
frequency that album art downloaders run it to get rid of the bug if it
affects spotify.